debug: isolate input provenance in its own extension
