+++ Project.toml
@@ -45,6 +45,7 @@ ReverseDiff = "37e2e3b7-166d-5795-8a7a-e32c996b4267"
 DynamicPPLComponentArraysExt = ["ComponentArrays"]
 DynamicPPLEnzymeCoreExt = ["EnzymeCore"]
 DynamicPPLForwardDiffExt = ["ForwardDiff"]
+DynamicPPLInputProvenanceExt = ["ForwardDiff"]
 DynamicPPLMCMCChainsExt = ["MCMCChains"]
 DynamicPPLMooncakeExt = ["Mooncake"]
 DynamicPPLMarginalLogDensitiesExt = ["MarginalLogDensities"]
